claims/frauddetection/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
oracledb==2.0.1
python-dateutil==2.8.2